+++ src/preload.js
@@ -7,10 +7,15 @@ contextBridge.exposeInMainWorld('electronAPI', {
   getDevices: () => ipcRenderer.invoke('get-devices'),
   saveDevices: (devices) => ipcRenderer.invoke('save-devices', devices),
   controlDevice: (deviceId, command) => ipcRenderer.invoke('control-device', deviceId, command),
+  getDeviceDetails: (deviceId) => ipcRenderer.invoke('get-device-details', deviceId),
+  getDeviceStatus: (deviceId) => ipcRenderer.invoke('get-device-status', deviceId),
   
   // User configuration
   getUserConfig: () => ipcRenderer.invoke('get-user-config'),
   saveUserConfig: (config) => ipcRenderer.invoke('save-user-config', config),
+  
+  // Connection testing
+  testConnection: () => ipcRenderer.invoke('test-connection'),
   
   // Event listeners
   onOpenPreferences: (callback) => ipcRenderer.on('open-preferences', callback)